% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoPhaseSPML_ord.R
\name{twoPhaseSPML_ord}
\alias{twoPhaseSPML_ord}
\title{Two-Phase Semiparametric Maximum Likelihood for Ordinal Outcomes}
\usage{
twoPhaseSPML_ord(
  formula,
  miscov,
  auxvar,
  family = NULL,
  data0,
  data1,
  start.values = NULL,
  verbose = FALSE,
  n_second,
  model_type,
  num_categories,
  N,
  osm = T,
  iteration = NULL,
  em_tol = 1e-05,
  max_iter = 1000,
  array_id = NULL
)
}
\arguments{
\item{formula}{The formula for the model.}

\item{miscov}{A formula specifying the missing covariates.}

\item{auxvar}{A formula specifying the auxiliary variables.}

\item{family}{The family of models to use. Defaults to NULL.}

\item{data0}{Data from the complement of Phase 2 data.}

\item{data1}{Data from Phase 2.}

\item{start.values}{Optional starting values for the parameters.}

\item{verbose}{Logical. If TRUE, prints detailed information during iterations. Defaults to FALSE.}

\item{n_second}{The number of Phase 2 samples.}

\item{model_type}{The type of ordinal model (e.g., "Cumulative Logit", "Adjacent Category", "Stopping Ratio").}

\item{num_categories}{The number of categories in the ordinal outcome.}

\item{N}{The total sample size.}

\item{osm}{Logical. If TRUE, uses the ordinal stereotype model by the OSM function otherwise it uses unordered stereotype with rrvglm. Defaults to TRUE.}

\item{iteration}{Optional iteration number.}

\item{em_tol}{The convergence tolerance for the EM algorithm. Defaults to 1e-05.}

\item{max_iter}{The maximum number of iterations for the EM algorithm. Defaults to 1000.}

\item{array_id}{Optional array ID for parallel processing.}
}
\value{
A list containing the estimated parameters, the probabilities for each category, the number of iterations, the log-likelihood, and the Fisher information matrix.
}
\description{
This function performs semiparametric maximum likelihood estimation for two-phase studies with ordinal outcomes using the EM algorithm.
}
\details{
This function estimates the parameters of a model for ordinal outcomes using data from a two-phase sampling design. The estimation is performed using the EM algorithm to handle the missing data and optimize the likelihood function. The function can handle different types of ordinal models including Cumulative Logit, Adjacent Category, and Stopping Ratio models.
}
\examples{
\dontrun{
data(biopsy)
# Example data preparation
data0 <- biopsy[1:300, ]
data1 <- biopsy[301:699, ]
formula <- ClumpThickness ~ CellShape + CellSize
miscov <- ~ CellShape
auxvar <- ~ CellSize
results <- twoPhaseSPML_ord(formula, miscov, auxvar, family=propodds(reverse=F),
                            data0, data1, n_second=100, model_type="Proportional_Odds",
                            num_categories=4, N=699)
}
}
